requirements.txt
pandas>=2.1.0
numpy>=1.26.0
ta==0.10.2
yfinance==0.2.18
pyTelegramBotAPI==4.12.0
pytz==2023.3
ratelimit==2.2.1
python-dotenv==1.0.0
